+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v1.5.0 (2024-03-02)
 
 * Fixed: Local UNC paths were only simplified for the C drive.
 * Added: `shawl --version` to display the program version.